requirements.txt
sqlalchemy
asyncpg
aiogram
pillow
python-dotenv